# [Numpy] FFI for argmax, argmin, indices